feat: Use real user ID for logged-in user management

+++ src/main/java/com/rofix/enotes_service/config/AppConfig.java
@@ -7,8 +7,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
 
 @Configuration
 @EnableJpaAuditing(auditorAwareRef = "auditorAware")
@@ -20,7 +18,7 @@ public class AppConfig {
     }
 
     @Bean
-    public AuditorAware<Integer> auditorAware()
+    public AuditorAware<Long> auditorAware()
     {
         return new AuditorAwareImpl();
     }
