+++ src/main/java/com/rofix/enotes_service/service/NoteService.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.rofix.enotes_service.dto.response.NoteResponseDTO;
 import com.rofix.enotes_service.dto.response.PageResponseDTO;
 import com.rofix.enotes_service.entity.FileDetails;
-import jakarta.validation.constraints.Min;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -17,7 +16,7 @@ public interface NoteService {
 
     byte[] downloadFile(FileDetails fileDetails) throws IOException;
 
-    PageResponseDTO getUserNotes(Integer userId, Integer pageNumber, Integer pageSize);
+    PageResponseDTO getUserNotes(Long userId, Integer pageNumber, Integer pageSize);
 
 
     NoteResponseDTO updateNote(Long id, String note, MultipartFile file) throws JsonProcessingException;
@@ -26,9 +25,9 @@ public interface NoteService {
 
     NoteResponseDTO restoreDeleteNote(Long id);
 
-    List<NoteResponseDTO> getUserRecycleBin(Integer userId);
+    List<NoteResponseDTO> getUserRecycleBin(Long userId);
 
-    String userClearRecycleBin(Integer userId);
+    String userClearRecycleBin(Long userId);
 
     void copyNote(Long noteId);
 }
